backend/requirements.txt
# Django Core
Django==5.2.7
djangorestframework==3.16.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# Database
psycopg2-binary==2.9.11

# Image Processing
Pillow==12.0.0

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.1.1

# Date/Time Handling
python-dateutil==2.9.0.post0
pytz==2025.2

# Celery & Task Queue
celery==5.4.0
redis==5.2.1
django-celery-beat==2.8.1
django-celery-results==2.6.0

# HTTP Requests (if needed for external APIs)
requests==2.32.5

# Essential dependencies (automatically included but listed for clarity)
asgiref==3.10.0
sqlparse==0.5.3
typing-extensions==4.15.0
